+++ DependencyInjection/ONGRElasticsearchExtension.php
@@ -13,15 +13,9 @@ namespace ONGR\ElasticsearchBundle\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\Loader;
-use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-use Symfony\Component\DependencyInjection\Reference;
 
-/**
- * This is the class that loads and manages bundle configuration.
- */
 class ONGRElasticsearchExtension extends Extension
 {
     /**
@@ -34,30 +28,17 @@ class ONGRElasticsearchExtension extends Extension
 
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.yml');
-        $config['cache'] = isset($config['cache']) ?
-            $config['cache'] : !$container->getParameter('kernel.debug');
-        $config['profiler'] = isset($config['profiler']) ?
-            $config['profiler'] : $container->getParameter('kernel.debug');
 
-        $container->setParameter('es.cache', $config['cache']);
-        $container->setParameter('es.analysis', $config['analysis']);
-        $container->setParameter('es.managers', $config['managers']);
-        $container->setParameter('es.indexes', $config['indexes']);
-        $definition = new Definition(
-            'ONGR\ElasticsearchBundle\Service\ManagerFactory',
-            [
-                new Reference('es.metadata_collector'),
-                new Reference('es.result_converter'),
-                $config['profiler'] ? new Reference('es.tracer') : null,
-            ]
+        $container->setParameter(
+            Configuration::ONGR_CACHE_CONFIG,
+            $config['cache'] ?? $container->getParameter('kernel.debug')
         );
-        $definition->addMethodCall('setEventDispatcher', [new Reference('event_dispatcher')]);
-        $definition->addMethodCall(
-            'setStopwatch',
-            [
-                new Reference('debug.stopwatch', ContainerInterface::NULL_ON_INVALID_REFERENCE)
-            ]
-        );
-        $container->setDefinition('es.manager_factory', $definition);
+
+        $container->setParameter(
+            Configuration::ONGR_PROFILER_CONFIG,
+            $config['profiler'] ?? $container->getParameter('kernel.debug')
+            );
+
+        $container->setParameter(Configuration::ONGR_ANALYSIS_CONFIG, $config['analysis']);
     }
 }