+++ ONGRElasticsearchBundle.php
@@ -11,23 +11,15 @@
 
 namespace ONGR\ElasticsearchBundle;
 
-use ONGR\ElasticsearchBundle\DependencyInjection\Compiler\ManagerFactoryPass;
 use ONGR\ElasticsearchBundle\DependencyInjection\Compiler\MappingPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-/**
- * ONGR Elasticsearch bundle system file required by kernel.
- */
 class ONGRElasticsearchBundle extends Bundle
 {
-    /**
-     * {@inheritdoc}
-     */
     public function build(ContainerBuilder $container)
     {
         parent::build($container);
-
         $container->addCompilerPass(new MappingPass());
     }
 }